Cleanup packet compression code
Just need to change add_buffer_to_vector to merge this
